unit_13/mysite/requirements/test.txt
# See more documentation in /docs/libraries.md

# Importing from base.txt
-r base.txt

# Measures code coverage
# http://nedbatchelder.com/code/coverage/
# coverage==3.7.1
coverage==4.0.2

# Easily specify directories to be excluded from testing
# https://github.com/kgrandis/nose-exclude/
nose-exclude==0.4.1

# Django test runner using nose
# https://github.com/django-nose/django-nose
django-nose==1.4.2

# Test fixtures replacement for Python
# https://github.com/rbarrois/factory_boy
factory-boy==2.6.0

# Python style checker
# http://flake8.readthedocs.org/en/latest/
flake8==2.5.0

django-autofixture==0.10.1

selenium==2.48.0

mock==1.3.0